refactor: refactor utilization package path
